requirements.txt
fastapi
uvicorn[standard]
celery
redis
pydantic
praw
google-cloud-texttospeech
nltk
mutagen
openai
moviepy